% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl}
\alias{tbl}
\title{Table function to create named variable summary list}
\usage{
tbl(df, vars)
}
\arguments{
\item{df}{Dataframe}

\item{vars}{Variable Selection (tidyselect)}
}
\value{
Named list of variable counts and percentages
}
\description{
Table function to create named variable summary list
}
\examples{
datafile <- system.file("penguins.RDS", package = "varSummary")
penguins <- readRDS(datafile)


tbl(penguins, starts_with("bill"))

tbl(penguins, ends_with("_mm"))

tbl(penguins, c(island, sex, species))
}
