% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSummary-package.R
\docType{package}
\name{varSummary-package}
\alias{varSummary}
\alias{varSummary-package}
\title{varSummary: Variable Summary}
\description{
Creating a named list with value count information of variables of a dataframe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/L-Groeninger/varSummary}
  \item Report bugs at \url{https://github.com/L-Groeninger/varSummary/issues}
}

}
\author{
\strong{Maintainer}: Lukas Gröninger \email{lukas.groeninger@gmail.com}

}
\keyword{internal}
